backend/requirements.txt
flask
gunicorn
requests
azure-identity
